Running paraview from python!
Added system paths to run paraview directly by running the .py file by pvpython. Running in linux looks like

 /home/oguz/Downloads/ParaView-5.5.2/bin/pvpython /path/to/code/Extract_Cells_Around_Sphere.py


input() at the end does not let the script finish so that the image stays visible. I do not know why without it the image disappears.(As the code is finished.)